refactor deleteImage. Check if file exists.

+++ src/controllers/deleteImage.js
@@ -1,12 +1,21 @@
 const db = require('../entities/Database');
+const { exists } = require('../utils/fs');
 const { NotFoundApiError } = require('../validators/errors/ApiError');
 
-module.exports = async (req, res) => {
+module.exports = async (req, res, next) => {
   const imageId = req.params.id;
+  const imageRaw = db.findOne(imageId);
   try {
+    const pathToFile = imageRaw.getImagePath(imageId);
+    const isFileExists = await exists(pathToFile);
+
+    if (!isFileExists) {
+      throw new NotFoundApiError('Image file not found');
+    }
+
     const id = await db.remove(imageId);
     return res.json({ id });
   } catch (err) {
-    throw new NotFoundApiError(`Image with ${imageId} doesn't exist`)
+    next(err);
   }
 };
